tiempo unix zona horaria

+++ Assets/Telemetria/Events/GameplayEvents.cs
@@ -9,7 +9,7 @@ public class Player_Death : TrackerEvent
     public float2 player_position = new float2(0, 0);
     public string cause = "";
 
-    public Player_Death(int timeStamp, string levelId, string roomId, float2 position, string cause) : base(timeStamp)
+    public Player_Death(long timeStamp, string levelId, string roomId, float2 position, string cause) : base(timeStamp)
     {
         this.level_id = levelId;
         this.room_id = roomId;
